# 🚀 Performance & Accessibility Optimization: Lighthouse Score 39→85+ Target