--- a/PERFORMANCE_OPTIMIZATION_SUMMARY.md
+++ b/PERFORMANCE_OPTIMIZATION_SUMMARY.md
@@ -1,0 +1,196 @@
+# Grade Spark Academy - Performance & Accessibility Optimization Summary
+
+## Overview
+This document summarizes the comprehensive performance and accessibility optimizations implemented to improve the website's Core Web Vitals scores and user experience.
+
+## Performance Optimizations (Target: 39 → 85+)
+
+### 1. Font Loading Optimization ✅
+**Issue**: Pacifico font blocking render, multiple Orbitron font loads
+**Solutions Implemented**:
+- Added `font-display: swap` to Pacifico font in `index.html`
+- Implemented font preloading with fallbacks for critical fonts (Inter, Orbitron)
+- Removed duplicate Orbitron font loading from GSA components
+- Added noscript fallbacks for browsers without preload support
+
+**Files Modified**:
+- `index.html` - Added preload links and font-display optimization
+- `src/components/GSAShieldLogo.tsx` - Removed duplicate font loading
+- `src/components/GSALogo.tsx` - Removed duplicate font loading
+- `src/components/common/GSAShieldLogo.tsx` - Removed duplicate font loading
+
+### 2. JavaScript Bundle Optimization ✅
+**Issue**: 67% waste in main.tsx, large unused bundles
+**Solutions Implemented**:
+- Enhanced code splitting in `vite.config.ts` with granular chunks
+- Lazy loaded performance monitoring to reduce initial bundle
+- Created lazy-loading wrapper for Three.js components
+- Separated vendor libraries into logical chunks (React, Three.js, Animation, UI, etc.)
+
+**Files Modified**:
+- `vite.config.ts` - Improved manual chunks configuration
+- `src/main.tsx` - Lazy loaded performance monitoring
+- `src/components/home/Hero.tsx` - Lazy loaded Three.js imports
+- `src/components/common/LazyThreeJS.tsx` - New lazy loading wrapper
+
+### 3. Text Compression ✅
+**Issue**: No Gzip/Brotli compression (12MB potential savings)
+**Solutions Implemented**:
+- Added `vite-plugin-compression` for both Gzip and Brotli
+- Configured automatic compression for production builds
+
+**Files Modified**:
+- `vite.config.ts` - Added compression plugins
+- `package.json` - Added compression dependency
+
+### 4. Cumulative Layout Shift (CLS) Fixes ✅
+**Issue**: Footer causing layout shifts (0.364 score)
+**Solutions Implemented**:
+- Added `min-h-[500px]` to footer to reserve space
+- Enhanced font loading to prevent FOUT/FOIT layout shifts
+
+**Files Modified**:
+- `src/components/Footer.tsx` - Added minimum height
+
+### 5. Performance Monitoring Enhancement ✅
+**Solutions Implemented**:
+- Enhanced Core Web Vitals tracking with PerformanceObserver
+- Added LCP, FID, and CLS monitoring
+- Improved long task detection
+
+**Files Modified**:
+- `src/lib/performance.ts` - Enhanced web vitals tracking
+
+## Accessibility Improvements (Target: 84 → 95+)
+
+### 1. Button Accessibility ✅
+**Issue**: Mobile nav toggle and FAQ buttons lack accessible names
+**Solutions Implemented**:
+- Added `aria-label`, `aria-expanded`, and `aria-controls` to mobile menu button
+- Added proper ARIA attributes to FAQ accordion buttons
+- Added `aria-hidden="true"` to decorative icons
+
+**Files Modified**:
+- `src/components/Navbar.tsx` - Enhanced mobile menu button accessibility
+- `src/components/home/FAQSection.tsx` - Enhanced FAQ button accessibility
+
+### 2. Link Accessibility ✅
+**Issue**: Logo links lack accessible names
+**Solutions Implemented**:
+- Added `aria-label="Grade Spark Academy Home"` to logo links
+
+**Files Modified**:
+- `src/components/Navbar.tsx` - Added logo link accessibility
+
+### 3. ARIA Attribute Fixes ✅
+**Issue**: Animated icons have `aria-label` without proper roles
+**Solutions Implemented**:
+- Added `role="img"` to animated icon divs with `aria-label`
+- Ensured proper ARIA usage throughout components
+
+**Files Modified**:
+- `src/components/home/Subjects.tsx` - Fixed icon ARIA roles
+
+### 4. Enhanced FAQ Accessibility ✅
+**Solutions Implemented**:
+- Added unique IDs to FAQ answers with `aria-controls` relationship
+- Added `role="region"` to FAQ answer containers
+- Implemented proper `aria-labelledby` relationships
+
+**Files Modified**:
+- `src/components/home/FAQSection.tsx` - Comprehensive FAQ accessibility
+
+## Best Practices & Diagnostics (Target: 74 → 90+)
+
+### 1. Console Error Fixes ✅
+**Issue**: Missing manifest icon and blocked Facebook Pixel
+**Solutions Implemented**:
+- Created comprehensive `manifest.json` with proper PWA configuration
+- Generated placeholder icon files to prevent 404 errors
+- Added manifest link to HTML head
+
+**Files Created/Modified**:
+- `public/manifest.json` - New PWA manifest
+- `public/icons/` - Generated placeholder icons
+- `index.html` - Added manifest link
+
+### 2. Testing Infrastructure ✅
+**Solutions Implemented**:
+- Created comprehensive accessibility test suite
+- Added tests for ARIA attributes, button accessibility, and link accessibility
+- Implemented performance-related accessibility tests
+
+**Files Created**:
+- `src/tests/accessibility.test.tsx` - Comprehensive accessibility tests
+
+## Technical Implementation Details
+
+### Code Splitting Strategy
+```typescript
+manualChunks: {
+  'vendor-react': ['react', 'react-dom'],
+  'vendor-router': ['react-router-dom'],
+  'vendor-three': ['three', '@react-three/fiber', '@react-three/drei', '@react-three/cannon'],
+  'vendor-animation': ['framer-motion', 'gsap'],
+  'vendor-ui': ['lucide-react', 'react-icons'],
+  'vendor-carousel': ['react-slick', 'swiper', 'slick-carousel'],
+  'vendor-utils': ['typewriter-effect', 'react-hot-toast', 'gray-matter'],
+  'vendor-analytics': ['@microsoft/clarity', '@supabase/supabase-js']
+}
+```
+
+### Font Loading Strategy
+```html
+<!-- Preload critical fonts -->
+<link rel="preload" href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700;800&display=swap" as="style">
+<link rel="preload" href="https://fonts.googleapis.com/css2?family=Orbitron:wght@400;700&display=swap" as="style">
+<link href="https://fonts.googleapis.com/css2?family=Pacifico&display=swap" rel="stylesheet">
+```
+
+### Accessibility Enhancements
+```typescript
+// Mobile menu button
+<button
+  aria-label={isOpen ? "Close navigation menu" : "Open navigation menu"}
+  aria-expanded={isOpen}
+  aria-controls="mobile-menu"
+>
+
+// FAQ buttons
+<button
+  aria-expanded={isOpen}
+  aria-controls={`faq-answer-${item.question.replace(/\s+/g, '-').toLowerCase()}`}
+  aria-label={`Toggle FAQ: ${item.question}`}
+>
+```
+
+## Expected Performance Improvements
+
+### Before Optimization
+- **Performance Score**: 39
+- **LCP**: 14.5 seconds
+- **CLS**: 0.364
+- **Accessibility Score**: 84
+- **Bundle Size**: 20.3 MB
+
+### After Optimization (Expected)
+- **Performance Score**: 85+ (Target)
+- **LCP**: <2.5 seconds (Good)
+- **CLS**: <0.1 (Good)
+- **Accessibility Score**: 95+ (Target)
+- **Bundle Size**: Reduced by ~40% through code splitting
+
+## Next Steps
+
+1. **Monitor Performance**: Use the enhanced performance monitoring to track improvements
+2. **Run Lighthouse**: Verify improvements with new Lighthouse audit
+3. **Test Accessibility**: Run automated accessibility tests
+4. **Optimize Images**: Consider implementing image optimization for further performance gains
+5. **CDN Implementation**: Consider CDN for static assets if not already implemented
+
+## Maintenance Notes
+
+- **Icon Files**: Replace placeholder icons with high-quality branded icons
+- **Font Loading**: Monitor font loading performance and adjust preload strategy as needed
+- **Bundle Analysis**: Regularly analyze bundle sizes and optimize as the application grows
+- **Accessibility Testing**: Include accessibility tests in CI/CD pipeline