--- a/src/main.tsx
+++ b/src/main.tsx
@@ -3,10 +3,15 @@ import { createRoot } from 'react-dom/client';
 import App from './App.tsx';
 import './styles/index.css';
 import ErrorBoundary from './components/common/ErrorBoundary';
-import { initPerformanceMonitoring } from './lib/performance';
 
-// Initialize performance monitoring
-initPerformanceMonitoring();
+// Lazy load performance monitoring to reduce initial bundle
+const initPerformanceMonitoring = async () => {
+  const { initPerformanceMonitoring } = await import('./lib/performance');
+  initPerformanceMonitoring();
+};
+
+// Initialize performance monitoring after initial render
+setTimeout(initPerformanceMonitoring, 100);
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>